Filtering and Using Arrow Function: Used an array method to filter the products and created another filteredProduct array container to contain a list of filtered products. Changed the table's ngfor structural directive to use the filteredProduct array container.
To specify the logic to filter our products we pass an arrow function into the filter method.

The this.products.filter() method iterates through each product in the array, passing each product into the arrow function. We have only one parameter (product: IProduct). We then use the arrow => to separate the parameter list from the function body.

A requirement is that the filtered list should only contain products with a product name that includes the listFilter (product.productName.includes(this.listFilter)).

This includes method .includes() returns true if the defined string is found in the product name, otherwise it returns false.